+++ packages/expo-router/src/__tests__/loadStaticParamsAsync.test.ios.ts
@@ -117,7 +117,7 @@ describe(loadStaticParamsAsync, () => {
           contextKey: './red.tsx',
           dynamic: null,
           route: 'red',
-
+          generated: true,
           entryPoints: ['expo-router/build/views/Navigator.js', './[color].tsx'],
         },
         {
@@ -126,6 +126,7 @@ describe(loadStaticParamsAsync, () => {
           contextKey: './blue.tsx',
           dynamic: null,
           route: 'blue',
+          generated: true,
           entryPoints: ['expo-router/build/views/Navigator.js', './[color].tsx'],
         },
       ],
@@ -227,6 +228,7 @@ describe(loadStaticParamsAsync, () => {
               dynamic: null,
               route: 'square',
               entryPoints: ['./_layout.tsx', './[color]/_layout.tsx', './[color]/[shape].tsx'],
+              generated: true,
             },
             {
               type: 'route',
@@ -235,6 +237,7 @@ describe(loadStaticParamsAsync, () => {
               dynamic: null,
               route: 'triangle',
               entryPoints: ['./_layout.tsx', './[color]/_layout.tsx', './[color]/[shape].tsx'],
+              generated: true,
             },
           ],
           type: 'layout',
@@ -242,6 +245,7 @@ describe(loadStaticParamsAsync, () => {
           dynamic: null,
           initialRouteName: undefined,
           route: 'red',
+          generated: true,
         },
         {
           children: [
@@ -260,6 +264,7 @@ describe(loadStaticParamsAsync, () => {
               dynamic: null,
               route: 'square',
               entryPoints: ['./_layout.tsx', './[color]/_layout.tsx', './[color]/[shape].tsx'],
+              generated: true,
             },
             {
               type: 'route',
@@ -268,6 +273,7 @@ describe(loadStaticParamsAsync, () => {
               dynamic: null,
               route: 'triangle',
               entryPoints: ['./_layout.tsx', './[color]/_layout.tsx', './[color]/[shape].tsx'],
+              generated: true,
             },
           ],
           type: 'layout',
@@ -275,6 +281,7 @@ describe(loadStaticParamsAsync, () => {
           dynamic: null,
           initialRouteName: undefined,
           route: 'blue',
+          generated: true,
         },
       ],
       type: 'layout',
@@ -439,6 +446,7 @@ describe(loadStaticParamsAsync, () => {
           dynamic: null,
           route: 'post/123/456',
           entryPoints: ['expo-router/build/views/Navigator.js', './post/[...post].tsx'],
+          generated: true,
         },
       ],
       type: 'layout',
@@ -534,6 +542,7 @@ describe(loadStaticParamsAsync, () => {
                 './(app)/(index,about)/blog/[post].tsx',
               ],
               route: '(index)/blog/123',
+              generated: true,
             },
             {
               type: 'route',
@@ -546,6 +555,7 @@ describe(loadStaticParamsAsync, () => {
                 './(app)/(index,about)/blog/[post].tsx',
               ],
               route: '(index)/blog/abc',
+              generated: true,
             },
             {
               type: 'route',
@@ -571,6 +581,7 @@ describe(loadStaticParamsAsync, () => {
               ],
 
               route: '(about)/blog/123',
+              generated: true,
             },
             {
               type: 'route',
@@ -583,6 +594,7 @@ describe(loadStaticParamsAsync, () => {
                 './(app)/(index,about)/blog/[post].tsx',
               ],
               route: '(about)/blog/abc',
+              generated: true,
             },
           ],
           type: 'layout',
@@ -672,6 +684,7 @@ describe(loadStaticParamsAsync, () => {
           children: [],
           contextKey: './post/123.tsx',
           dynamic: null,
+          generated: true,
           route: 'post/123',
           entryPoints: ['expo-router/build/views/Navigator.js', './post/[post].tsx'],
         },
@@ -701,6 +714,7 @@ describe(loadStaticParamsAsync, () => {
           children: [],
           contextKey: './a/b/c/d/e.tsx',
           dynamic: null,
+          generated: true,
           route: 'a/b/c/d/e',
           entryPoints: ['expo-router/build/views/Navigator.js', './a/[b]/c/[d]/[e].tsx'],
         },